@@ -1,6 +1,6 @@
 {
   "name": "react-sparklines",
-  "version": "1.5.1",
+  "version": "1.5.2",
   "description": "Beautiful and expressive Sparklines component for React ",
   "main": "build/index.js",
   "directories": {
@@ -35,7 +35,7 @@
   "homepage": "https://github.com/borisyankov/react-sparklines#readme",
   "devDependencies": {
     "babel": "^6.5.2",
-    "babel-core": "^6.7.2",
+    "babel-core": "^6.7.6",
     "babel-loader": "^6.2.4",
     "babel-plugin-transform-object-assign": "^6.5.0",
     "babel-preset-es2015": "^6.6.0",
@@ -43,20 +43,20 @@
     "babel-preset-stage-1": "^6.5.0",
     "babel-runtime": "^6.6.1",
     "chai": "^3.5.0",
-    "enzyme": "^2.1.0",
+    "enzyme": "^2.2.0",
     "hiff": "^0.3.0",
     "line-by-line": "^0.1.4",
     "mocha": "^2.4.5",
-    "react": "^0.14.7",
-    "react-addons-test-utils": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "react": "^15.0.1",
+    "react-addons-test-utils": "^15.0.1",
+    "react-dom": "^15.0.1",
     "react-element-to-jsx-string": "=2.5.0",
     "replaceall": "^0.1.6",
     "webpack": "^2.1.0-beta.4",
     "webpack-dev-server": "^2.0.0-beta"
   },
   "peerDependencies": {
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7"
+    "react": "*",
+    "react-dom": "*"
   }
 }
